+++ SolidworksAddinFramework/EventHandlers/OpenGL.cs
@@ -7,7 +7,7 @@ public static class OpenGl
     [DllImport("opengl32")]
     public static extern void glEnd();
     [DllImport("opengl32")]
-    public static extern void glVertex3f(float x, float y, float z);
+    public static extern void glVertex3f(double x, double y, double z);
 
     [DllImport("opengl32")]
     public static extern void glLineWidth(int width);
