+++ SolidworksAddinFramework/Geometry/FaceExtensions.cs
@@ -10,8 +10,8 @@ namespace SolidworksAddinFramework.Geometry
         {
             var box = (double[]) face.GetBox();
             return new Range3Single(
-                (float) box[0], (float) box[1], (float) box[2], 
-                (float) box[3], (float) box[4], (float) box[5]);
+                (double) box[0], (double) box[1], (double) box[2], 
+                (double) box[3], (double) box[4], (double) box[5]);
         }
 
         public static bool GetDistance(this IFace2 entity0, IFace2 entity1, out double[] posacast, out double[] posbcast)
@@ -28,9 +28,9 @@ namespace SolidworksAddinFramework.Geometry
             return result == 0;
         }
 
-        private static float[][] GetTessTrianglesTs(this IFace2 face, bool noConversion)
+        private static double[][] GetTessTrianglesTs(this IFace2 face, bool noConversion)
         {
-            var data = (float[])face.GetTessTriangles(noConversion);
+            var data = (double[])face.GetTessTriangles(noConversion);
             return data.Select((value, index) => new { value, index })
                 .GroupBy(x => x.index / 3, x => x.value) // a group is a point of the triangle
                 .Select(x => x.ToArray())
@@ -53,8 +53,8 @@ namespace SolidworksAddinFramework.Geometry
                 .ToList();
         }
 
-        public static float[][] GetTessTrianglesNoConversion(this IFace2 face) => face.GetTessTrianglesTs(true);
-        public static float[][] GetTessTrianglesAllowConversion(this IFace2 face) => face.GetTessTrianglesTs(false);
+        public static double[][] GetTessTrianglesNoConversion(this IFace2 face) => face.GetTessTrianglesTs(true);
+        public static double[][] GetTessTrianglesAllowConversion(this IFace2 face) => face.GetTessTrianglesTs(false);
 
         /// <summary>
         /// Packs trimming loops into an array suitable for usage by ISurface::CreateTrimmedSheet4
