+++ SolidworksAddinFramework/OpenGl/IRenderable.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Numerics;
+using System.DoubleNumerics;
 using LanguageExt;
 using SolidWorks.Interop.sldworks;
 using OpenTK.Graphics.OpenGL;
@@ -12,12 +12,12 @@ namespace SolidworksAddinFramework.OpenGl
     {
         void Render(DateTime time);
         void ApplyTransform(Matrix4x4 transform);
-        Tuple<Vector3, float> BoundingSphere { get; }
+        Tuple<Vector3, double> BoundingSphere { get; }
     }
 
     public static class Renderable
     {
-        public static Tuple<Vector3,float> BoundingSphere(IReadOnlyList<Vector3> points)
+        public static Tuple<Vector3,double> BoundingSphere(IReadOnlyList<Vector3> points)
         {
             var range = Range3Single.FromVertices(points);
             return range.BoundingSphere();
